Remove attempt to calculate number of parallel lines in case of voltage issues
As the voltage at the LV busbar changes significantly due to
reinforcement in other LV grids with voltage issues the voltage at the
LV busbar cannot be considered constant as is currently done in the
calculation of needed parallel lines. The calculated number of parallel
lines is therefore extremely overestimated... A new rule of thumb has to
be found.