Fix missing time series of status quo generator
Sometimes there was no time series available for wind power plants.
Based on the assumption that this happens when for example the grid
include wind generators but only in "Status quo" scenario, I changed the
feedin time series importer to take take from both scenarios "status
quo" and "nep 2035". If both is available, values are averaged.